chore: update from review
